// api/src/interfaces/customerAttributes.ts
export interface CustomerAttributes {
    CustomerID: string;
    CustomerName: string;
    Gender?: string;
    DateofBirth?: Date;
    PhoneNumber?: string;
    PackID?: string;
    StaffID?: string;
    Startdate?: Date;
    DateRegister?: number;
    Note?: string;
    Date_Added?: Date;
    Date_Modified?: Date;
    Image: string;
}